PIM-6118: Add behat for mass edit families available attribute button
