+++ features/Context/Page/Base/Form.php
@@ -7,6 +7,7 @@ use Behat\Mink\Element\ElementInterface;
 use Behat\Mink\Element\NodeElement;
 use Behat\Mink\Exception\ElementNotFoundException;
 use Behat\Mink\Exception\ExpectationException;
+use Context\Spin\TimeoutException;
 use Context\Traits\ClosestTrait;
 use Pim\Behat\Decorator\Field\Select2Decorator;
 
@@ -858,5 +859,21 @@ class Form extends Base
 
             return $tabs;
         }, 'Cannot find any tabs in this page');
+    }
+
+    /**
+     * Returns if the "add available attributes" button is enabled
+     *
+     * @return bool
+     *
+     * @throws TimeoutException
+     */
+    public function isAvailableAttributeEnabled()
+    {
+        $button = $this->spin(function () {
+            return $this->find('css', $this->elements['Available attributes button']['css']);
+        }, 'Cannot find available attribute button');
+        
+        return !$this->getClosest($button, 'select2-container')->hasClass('select2-container-disabled');
     }
 }
